+++ functions/index.js
@@ -1,7 +1,11 @@
 const pick = require('./pick');
+const resolvePaymentAction = require('./resolve-payment-action');
 const validateField = require('./validate-field');
+const verifyNotification = require('./verify-notification');
 
 module.exports = {
   pick,
+  resolvePaymentAction,
   validateField,
+  verifyNotification,
 };