+++ middlewares/index.js
@@ -1,5 +1,7 @@
+const captureRawBody = require('./capture-raw-body');
 const rawTextBodyParser = require('./raw-text-body-parser');
 
 module.exports = {
+  captureRawBody,
   rawTextBodyParser,
 };